Refactor frequency view into separate file
